Remove Venue dependency and booking concerns from VenueExpression
Booking seats in VenueExpression breaks the free ordering in expression
and the terminal ActionExpression context is guaranteed to have all
expression values parsed.
The base implementation of the venue dependency is a bit off compared
to the rest of the design. Will try with offering a func/lambda to
resolve the venue.